+++ src/Router/DispatchResult.php
@@ -5,26 +5,32 @@ namespace QuickRoute\Router;
 
 
 use FastRoute\Dispatcher as FastDispatcher;
+use InvalidArgumentException;
 
 class DispatchResult
 {
     /**
-     * @var mixed[]
+     * @var array
      */
     private array $dispatchResult;
 
+    private array $collectedRoutes;
+
     /**
      * DispatchResult constructor.
+     *
      * @param string[] $dispatchResult
      */
-    public function __construct(array $dispatchResult)
+    public function __construct(array $dispatchResult, array $collectedRoutes)
     {
         $this->dispatchResult = $dispatchResult;
+        $this->collectedRoutes = $collectedRoutes;
     }
 
 
     /**
      * If url is found
+     *
      * @return bool
      */
     public function isFound(): bool
@@ -34,6 +40,7 @@ class DispatchResult
 
     /**
      * If url is not found
+     *
      * @return bool
      */
     public function isNotFound(): bool
@@ -43,6 +50,7 @@ class DispatchResult
 
     /**
      * If url method is not allowed
+     *
      * @return bool
      */
     public function isMethodNotAllowed(): bool
@@ -52,7 +60,7 @@ class DispatchResult
 
     /**
      * Get dispatched url parameters
-     * @return array<mixed>|null
+     * @return array|null
      */
     public function getUrlParameters(): ?array
     {
@@ -60,7 +68,68 @@ class DispatchResult
     }
 
     /**
-     * Get found url class
+     * Find route by name
+     *
+     * @param string $routeName
+     * @return array|null
+     */
+    public function route(string $routeName): ?array
+    {
+        foreach ($this->collectedRoutes as $collectedRoute) {
+            if ($routeName == $collectedRoute['name']) {
+                return $collectedRoute;
+            }
+        }
+
+        return null;
+    }
+
+    /**
+     * Generate route http uri using route's name
+     *
+     * @param string $routeName
+     * @param array $routeParams an array of [key => value] of route parameters
+     * @return string|null
+     */
+    public function uri(string $routeName, array $routeParams = []): ?string
+    {
+        $foundRoute = $this->route($routeName);
+        if (!$foundRoute) return null;
+
+        $prefix = $foundRoute['prefix'] ?? null;
+        if (!$prefix) return null;
+
+        return $this->replaceParamWithValue($prefix, $routeParams);
+    }
+
+    protected function replaceParamWithValue(string $prefix, array $params): string
+    {
+        preg_match_all("@{([0-9a-zA-Z]+):?.*?\+?}@", $prefix, $matchedParams);
+
+        for ($i = 0; $i < count($matchedParams[0]); $i++) {
+            $paramValue = $params[$matchedParams[1][$i]] ?? null;
+            if (null == $paramValue) {
+                throw new InvalidArgumentException("Missing route parameter value for \"{$matchedParams[1][$i]}\"");
+            }
+            $prefix = str_replace($matchedParams[0][$i], $params[$matchedParams[1][$i]], $prefix);
+        }
+
+        return $prefix;
+    }
+
+    /**
+     * Get all collected routes
+     *
+     * @return array
+     */
+    public function getCollectedRoutes(): array
+    {
+        return $this->collectedRoutes;
+    }
+
+    /**
+     * Get found url
+     *
      * @return RouteData
      */
     public function getRoute(): RouteData